@@ -1,6 +1,4 @@
-from __future__ import unicode_literals
 from django.db import models
-from django.utils.encoding import python_2_unicode_compatible
 from uuid import uuid4
 
 import mptt
@@ -22,7 +20,6 @@ class CustomTreeManager(TreeManager):
         return CustomTreeQueryset(model=self.model, using=self._db)
 
 
-@python_2_unicode_compatible
 class Category(MPTTModel):
     name = models.CharField(max_length=50)
     parent = TreeForeignKey(
@@ -38,7 +35,6 @@ class Category(MPTTModel):
     delete.alters_data = True
 
 
-@python_2_unicode_compatible
 class Item(models.Model):
 
     name = models.CharField(max_length=100)
@@ -53,7 +49,6 @@ class Item(models.Model):
         return self.name
 
 
-@python_2_unicode_compatible
 class Genre(MPTTModel):
     name = models.CharField(max_length=50, unique=True)
     parent = TreeForeignKey(
@@ -79,7 +74,6 @@ class Insert(MPTTModel):
         on_delete=models.CASCADE)
 
 
-@python_2_unicode_compatible
 class MultiOrder(MPTTModel):
     name = models.CharField(max_length=50)
     size = models.PositiveIntegerField()
@@ -120,7 +114,6 @@ class UUIDNode(MPTTModel):
         return self.name
 
 
-@python_2_unicode_compatible
 class OrderedInsertion(MPTTModel):
     name = models.CharField(max_length=50)
     parent = TreeForeignKey(
@@ -145,11 +138,10 @@ class NewStyleMPTTMeta(MPTTModel):
         'self', null=True, blank=True, related_name='children',
         on_delete=models.CASCADE)
 
-    class MPTTMeta(object):
+    class MPTTMeta:
         left_attr = 'testing'
 
 
-@python_2_unicode_compatible
 class Person(MPTTModel):
     name = models.CharField(max_length=50)
     parent = TreeForeignKey(
@@ -167,7 +159,6 @@ class Student(Person):
     type = models.CharField(max_length=50)
 
 
-@python_2_unicode_compatible
 class CustomPKName(MPTTModel):
     my_id = models.AutoField(db_column='my_custom_name', primary_key=True)
     name = models.CharField(max_length=50)
